Point and Tier

+++ lib/object/tier.dart
@@ -1,7 +1,7 @@
 class Tier {
   int id;
   int loyaltyProgramId;
-  double minPoints;
+  int minPoints;
   double ratioPoints;
   int sequenceNumber;
   Tier(this.id, this.loyaltyProgramId, this.minPoints, this.ratioPoints,
